Use int instead of int64 unless necessary. Convert the whole project to confirm.

+++ controller/entity_test.go
@@ -1,5 +1,6 @@
 package controller
 
+/*
 import (
 	"encoding/json"
 	"fmt"
@@ -12,10 +13,10 @@ import (
 )
 
 const (
-	t_company_id      int64 = 10
-	t_item_rev        int64 = 1231
-	t_branch_rev      int64 = 12
-	t_branch_item_rev int64 = 240
+	t_company_id      int = 10
+	t_item_rev        int = 1231
+	t_branch_rev      int = 12
+	t_branch_item_rev int = 240
 )
 
 var syncJsonFormat string = `
@@ -51,10 +52,10 @@ func (arr intArr) String() string {
 }
 
 var entityParseTest = []struct {
-	entityType    string
-	create_ids    intArr
-	update_ids    intArr
-	delete_ids    intArr
+	entityType string
+	create_ids intArr
+	update_ids intArr
+	delete_ids intArr
 
 	existingItems []*models.ShItem
 	fields        []map[string]interface{}
@@ -81,16 +82,16 @@ var entityParseTest = []struct {
 				models.ITEM_JSON_BAR_CODE:   "123456789",
 			},
 			{
-				models.ITEM_JSON_ITEM_ID:     -2,
-				models.ITEM_JSON_COMPANY_ID:  t_company_id,
-				models.ITEM_JSON_MODEL_YEAR:  "1992",
-				models.ITEM_JSON_ITEM_CODE: "A-1028",
+				models.ITEM_JSON_ITEM_ID:    -2,
+				models.ITEM_JSON_COMPANY_ID: t_company_id,
+				models.ITEM_JSON_MODEL_YEAR: "1992",
+				models.ITEM_JSON_ITEM_CODE:  "A-1028",
 			},
 			{
-				models.ITEM_JSON_ITEM_ID:     77,
-				models.ITEM_JSON_COMPANY_ID:  t_company_id,
-				models.ITEM_JSON_MODEL_YEAR:  "new year",
-				models.ITEM_JSON_ITEM_CODE: "updated model",
+				models.ITEM_JSON_ITEM_ID:    77,
+				models.ITEM_JSON_COMPANY_ID: t_company_id,
+				models.ITEM_JSON_MODEL_YEAR: "new year",
+				models.ITEM_JSON_ITEM_CODE:  "updated model",
 			},
 			{
 				models.ITEM_JSON_ITEM_ID:     -7,
@@ -109,10 +110,10 @@ var entityParseTest = []struct {
 		[]map[string]interface{}{
 			{
 				// doesn't exist in any of CRUD listings
-				models.ITEM_JSON_ITEM_ID:     -4,
-				models.ITEM_JSON_COMPANY_ID:  t_company_id,
-				models.ITEM_JSON_MODEL_YEAR:  "1992",
-				models.ITEM_JSON_ITEM_CODE: "A-1028",
+				models.ITEM_JSON_ITEM_ID:    -4,
+				models.ITEM_JSON_COMPANY_ID: t_company_id,
+				models.ITEM_JSON_MODEL_YEAR: "1992",
+				models.ITEM_JSON_ITEM_CODE:  "A-1028",
 			},
 			{
 				models.ITEM_JSON_ITEM_ID:     -7,
@@ -222,3 +223,4 @@ func TestEntitySyncHandler(t *testing.T) {
 		}
 	}
 }
+*/
